Небольшой рефактор кода. Была удеална вермия для jpa data, был отредактирован dog класс, были пересены все сервисы в одну папку

+++ src/main/java/animal/shelter/animalsshelter/service/UserService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface UserService {
     User saveUser(User user);
-    User getUserById(long id);
+    User getUserById(Integer id);
     List<User> getAllUsers();
-    void deleteUser(long id);
+    void deleteUser(Integer id);
 }
